+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/IntelliJPlatformType.kt
@@ -215,8 +215,14 @@ enum class IntelliJPlatformType(
  */
 @Throws(IllegalArgumentException::class)
 fun Any.toIntelliJPlatformType(version: String) = when (this) {
-    is IntelliJPlatformType -> IntelliJPlatformType.fromCode(code, version)
-    is String -> IntelliJPlatformType.fromCode(this, version)
+    is IntelliJPlatformType -> when (version) {
+        Constraints.LATEST_VERSION -> this
+        else -> IntelliJPlatformType.fromCode(code, version)
+    }
+    is String -> when (version) {
+        Constraints.LATEST_VERSION -> IntelliJPlatformType.fromCode(this)
+        else -> IntelliJPlatformType.fromCode(this, version)
+    }
     else -> throw IllegalArgumentException("Invalid argument type: '$javaClass'. Supported types: String or ${IntelliJPlatformType::class.java}")
 }